+++ src/main/java/musicclient/endpoint/SyncEndpoint.java
@@ -5,12 +5,12 @@ import music.model.Track;
 import musicclient.model.impl.SyncResult;
 import musicclient.service.SyncService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.env.Environment;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.util.List;
 
@@ -21,17 +21,17 @@ public class SyncEndpoint {
     private final SyncService syncService;
 
     @Autowired
+    Environment environment;
+
+    @Autowired
     public SyncEndpoint(SyncService syncService) {
         this.syncService = syncService;
     }
 
     @PostMapping
-    public SyncResult sample(@RequestBody List<Track> tracksToSync, HttpServletRequest request) throws IOException, TaskInProgressException {
-        String serverName = request.getServerName();
-        int serverPort = request.getServerPort();
-        String scheme = request.getScheme();
-        String contextPath = request.getContextPath();
+    public SyncResult sample(@RequestBody List<Track> tracksToSync) throws IOException, TaskInProgressException {
+        String serverPort = environment.getProperty("local.server.port");
 
-        return syncService.performSync(tracksToSync, serverName, serverPort, scheme, contextPath);
+        return syncService.performSync(tracksToSync, serverPort);
     }
 }